adjust padding and fontsize and util
