+++ src/components/layout/index.module.scss
@@ -2,7 +2,7 @@
   width: 100vw;
   height: 100vh;
   background: var(--primary-theme-color);
-  padding: 20px 24px 0 0;
+  padding: 8px 5px 0 0;
   display: flex;
 
   .left {
