+++ src/pages/charts/components/data-panel/index.module.scss
@@ -60,7 +60,7 @@
 
   .echartsPanel {
     .topInfo {
-      height: 60px;
+      height: 52px;
       padding: 0 24px;
       background: var(--primary-content-color);
       display: flex;
