+++ src/pages/charts/components/header/index.module.scss
@@ -55,7 +55,7 @@
       flex-shrink: 0;
       .title {
         font-weight: 400;
-        font-size: 12px;
+        font-size: 10px;
         line-height: 12px;
         color: #ffffff;
         opacity: 0.45;
@@ -65,7 +65,7 @@
       .valueInfo {
         span {
           font-weight: bold;
-          font-size: 14px;
+          font-size: 12px;
           line-height: 14px;
           opacity: 0.85;
           font-family: "DIN Alternate";
