+++ src/pages/charts/components/search/index.module.scss
@@ -4,7 +4,7 @@
 
   .filter {
     background: var(--primary-content-color);
-    padding: 16px;
+    padding: 12px;
   }
 
   .category {
