+++ src/pages/charts/index.module.scss
@@ -10,7 +10,7 @@
   .content {
     flex: 1;
     color: #fff;
-    margin-top: 12px;
+    margin-top: 4px;
     display: flex;
   }
 }
