+++ src/utils/echarts/independent-line.ts
@@ -1,6 +1,8 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
 // @ts-nocheck
-import { getPriceSeries,getIndenpendYAxis,getToolTipFormater, commonOption, padArrayAhead,formatBigNumber } from "./common";
+import { getPriceSeries,getIndenpendYAxis,getToolTipFormater, commonOption, padArrayAhead } from "./common";
+import {formatBigNumber} from "@/utils/common";
+
 
 export function independentLineTransform({
   indicatorData,
